backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
websockets==13.1
azure-identity==1.19.0
azure-monitor-query==1.4.0
openai==1.52.0
PyGithub==2.4.0
python-multipart==0.0.12
